+++ src/index.js
@@ -1,44 +1,44 @@
-import { readFile } from 'fs'
-import { createServer } from 'http'
-import { resolve } from 'path'
+import { readFile } from 'fs';
+import { createServer as createHttpsServer } from 'https';
+import { createServer } from 'http';
+import { resolve } from 'path';
 
-import mime from 'mime'
-import opener from 'opener'
+import mime from 'mime';
+import opener from 'opener';
 
-export default function serve (options = { contentBase: '' }) {
+export default function serve(options = { contentBase: '' }) {
   if (Array.isArray(options) || typeof options === 'string') {
-    options = { contentBase: options }
+    options = { contentBase: options };
   }
-  options.contentBase = Array.isArray(options.contentBase) ? options.contentBase : [options.contentBase]
-  options.host = options.host || 'localhost'
-  options.port = options.port || 10001
-  options.headers = options.headers || [];
+  options.contentBase = Array.isArray(options.contentBase) ? options.contentBase : [options.contentBase];
+  options.host = options.host || 'localhost';
+  options.port = options.port || 10001;
+  options.headers = options.headers || {};
+  options.https = options.https || false;
+  mime.default_type = 'text/plain';
 
-  mime.default_type = 'text/plain'
-
-  createServer(function (request, response) {
+  // Fallback to http protocol if https option is false or SSL cert files do not exist
+  const PROTOCOL = (options.https && 'https://') || 'http://';
+  const requestListener = (request, response) => {
     // Remove querystring
-    const urlPath = decodeURI(request.url.split('?')[0])
+    const urlPath = decodeURI(request.url.split('?')[0]);
 
-    options.headers.map((header) => {
-      response.setHeader(header.name, header.value);
+    Object.keys(options.headers).forEach((key) => {
+      response.setHeader(key, options.headers[key]);
     });
 
-    readFileFromContentBase(options.contentBase, urlPath, function (error, content, filePath) {
-      if (!error)  {
+    readFileFromContentBase(options.contentBase, urlPath, (error, content, filePath) => {
+      if (!error) {
         return found(response, filePath, content)
       }
       if (error.code !== 'ENOENT') {
         response.writeHead(500)
-        response.end('500 Internal Server Error' +
-          '\n\n' + filePath +
-          '\n\n' + Object.keys(error).map(function (k) { return error[k]; }).join('\n') +
-          '\n\n(rollup-plugin-serve)', 'utf-8')
+        response.end(`500 Internal Server Error\n\n${filePath}\n\n${Object.keys(error).map(k => error[k]).join('\n')}\n\n(rollup-plugin-serve)`, 'utf-8')
         return
       }
       if (request.url === '/favicon.ico') {
         filePath = resolve(__dirname, '../dist/favicon.ico')
-        readFile(filePath, function (error, content) {
+        readFile(filePath, (error, content) => {
           if (error) {
             notFound(response, filePath)
           } else {
@@ -46,7 +46,7 @@ export default function serve (options = { contentBase: '' }) {
           }
         })
       } else if (options.historyApiFallback) {
-        readFileFromContentBase(options.contentBase, '/index.html', function (error, content, filePath) {
+        readFileFromContentBase(options.contentBase, '/index.html', (error, content, filePath) => {
           if (error) {
             notFound(response, filePath)
           } else {
@@ -57,20 +57,26 @@ export default function serve (options = { contentBase: '' }) {
         notFound(response, filePath)
       }
     })
-  }).listen(options.port)
+  };
 
-  var running = options.verbose === false
+  if (options.https) {
+    createHttpsServer(options.https, requestListener).listen(options.port);
+  } else {
+    createServer(requestListener).listen(options.port);
+  }
+
+  let running = options.verbose === false
 
   return {
     name: 'serve',
-    ongenerate () {
+    ongenerate() {
       if (!running) {
         running = true
 
         // Log which url to visit
-        const url = 'http://' + options.host + ':' + options.port
+        const url = `${PROTOCOL + options.host}:${options.port}`
         options.contentBase.forEach(base => {
-          console.log(green(url) + ' -> ' + resolve(base))
+          console.log(`${green(url)} -> ${resolve(base)}`)
         })
 
         // Open browser
@@ -82,8 +88,8 @@ export default function serve (options = { contentBase: '' }) {
   }
 }
 
-function readFileFromContentBase (contentBase, urlPath, callback) {
-  let filePath = resolve(contentBase[0] || '.', '.' + urlPath)
+function readFileFromContentBase(contentBase, urlPath, callback) {
+  let filePath = resolve(contentBase[0] || '.', `.${urlPath}`)
 
   // Load index.html in directories
   if (urlPath.endsWith('/')) {
@@ -101,18 +107,16 @@ function readFileFromContentBase (contentBase, urlPath, callback) {
   })
 }
 
-function notFound (response, filePath) {
+function notFound(response, filePath) {
   response.writeHead(404)
-  response.end('404 Not Found' +
-    '\n\n' + filePath +
-    '\n\n(rollup-plugin-serve)', 'utf-8')
+  response.end(`404 Not Found\n\n${filePath}\n\n(rollup-plugin-serve)`, 'utf-8')
 }
 
-function found (response, filePath, content) {
+function found(response, filePath, content) {
   response.writeHead(200, { 'Content-Type': mime.lookup(filePath) })
   response.end(content, 'utf-8')
 }
 
-function green (text) {
-  return '\u001b[1m\u001b[32m' + text + '\u001b[39m\u001b[22m'
+function green(text) {
+  return `\u001b[1m\u001b[32m${text}\u001b[39m\u001b[22m`
 }